+++ lightstep/sdk/metric/exporters/otlp/otelcol/client_test.go
@@ -27,9 +27,10 @@ import (
 	sdkmetric "github.com/lightstep/otel-launcher-go/lightstep/sdk/metric"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/aggregator/aggregation"
+	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/internal/export"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/number"
 	"github.com/lightstep/otel-launcher-go/lightstep/sdk/metric/view"
-	"github.com/open-telemetry/otel-arrow/collector/receiver/otelarrowreceiver"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/otelarrowreceiver"
 	"github.com/stretchr/testify/suite"
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config/confignet"
@@ -185,7 +186,7 @@ func (t *clientTestSuite) SetupSuite() {
 	cfg.Protocols.Arrow = otelarrowreceiver.ArrowConfig{}
 	cfg.GRPC.NetAddr = confignet.AddrConfig{Endpoint: t.addr, Transport: "tcp"}
 
-	set := receivertest.NewNopCreateSettings()
+	set := receivertest.NewNopSettings()
 	tc := &consumertest.MetricsSink{}
 
 	mr, err := factory.CreateMetricsReceiver(ctx, set, cfg, tc)
@@ -651,7 +652,7 @@ func (t *clientTestSuite) TestFilteredAttributes() {
 				attrs := attribute.NewSet(useAttrs...)
 
 				dest := pmetric.NewExemplarSlice()
-				copyExemplars(
+				export.CopyExemplars(
 					dest, attrs, nk,
 					[]aggregator.WeightedExemplarBits{
 						{