+++ lightstep/sdk/metric/internal/viewstate/viewstate.go
@@ -64,7 +64,7 @@ type Compiler struct {
 
 	// library is the value used fr matching
 	// instrumentation library information.
-	library instrumentation.Library
+	library instrumentation.Scope
 
 	// lock protects collectors and names.
 	compilerLock sync.Mutex
@@ -189,7 +189,7 @@ type singleBehavior struct {
 }
 
 // New returns a compiler for library given configured views.
-func New(library instrumentation.Library, views *view.Views) *Compiler {
+func New(library instrumentation.Scope, views *view.Views) *Compiler {
 	views, _ = view.Validate(views)
 	return &Compiler{
 		library: library,
@@ -532,7 +532,6 @@ func compileSync[N number.Any, Traits number.Traits[N]](behavior singleBehavior)
 			sum.Methods[N, Traits, sum.NonMonotonic],
 		](behavior)
 	case aggregation.GaugeKind:
-		// Note: off-spec synchronous gauge support
 		return newSyncViewWithEx[
 			N,
 			Traits,
@@ -739,6 +738,15 @@ func checkSemanticCompatibility(ik sdkinstrument.Kind, behavior *singleBehavior)
 
 	case sdkinstrument.AsyncGauge:
 		switch cat {
+		case aggregation.GaugeCategory:
+			return nil
+		}
+
+	case sdkinstrument.SyncGauge:
+		switch cat {
+		// TODO: consider letting gauges have semantic compatibility
+		// with histogram aggregation, histograms have semantic compat
+		// with gauge aggregation, and so on.
 		case aggregation.GaugeCategory:
 			return nil
 		}